Move CRUD logic from palette app to Vue mixin
